+++ src/commands/CL02ledSetName.h
@@ -1,6 +1,6 @@
 void cL02ledSetName(){
     // Set name to Led
-    int lednr = commandParts.arg1.toInt();
+    size_t lednr = commandParts.arg1.toInt();
     Println("L02 Set name of " + commandParts.arg1 + " to " + commandParts.arg2);
 
     if (lednr >= 0 && lednr < LED_NR_ITEMS) {
