+++ src/commands/cK04kringEvery.h
@@ -1,7 +1,7 @@
 void cK04kringEvery(){
     // Set Every for Kring
-    int kringnr = commandParts.arg1.toInt();
-    int every = commandParts.arg2.toInt();
+    size_t kringnr = commandParts.arg1.toInt();
+    size_t every = commandParts.arg2.toInt();
     if (kringnr >= 0 && kringnr < RGB_NR_ITEMS) {
         kring[kringnr].every = every;
     } else {
