+++ src/led/led_uit.h
@@ -1,5 +1,5 @@
 #pragma once
 
-void led_uit(int lednr) {
+void led_uit(size_t lednr) {
     ledPCA9685.setPWM(lednr, 0, 0);
 }
